# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "contact": {
      "name": "CoE-DSC",
      "url": "https://coe-dsc.nl/"
    },
    "description": "API for the MAMI Framework DSI Assessment Tool \u2014 CoE-DSC / TNO",
    "title": "MAMI Checker API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/initiatives/{initiative_id}/report": {
      "get": {
        "description": "Retrieve the stored compliance report HTML for an initiative.",
        "operationId": "get_report_api_v1_initiatives__initiative_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Report",
        "tags": [
          "reports"
        ]
      },
      "post": {
        "description": "Generate a compliance report for an initiative.\n\nScores all saved answers, loads evidence, renders a full HTML report via\nJinja2, and upserts the result into the compliance_report table.\nReturns the rendered HTML directly.",
        "operationId": "generate_report_api_v1_initiatives__initiative_id__report_post",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Generate Report",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/initiatives/{initiative_id}/report/data": {
      "get": {
        "description": "Retrieve structured JSON report data for an initiative.\n\nRe-runs scoring on the fly from stored answers to avoid storing JSON separately.\nReturns 404 if no report has been generated yet.",
        "operationId": "get_report_data_endpoint_api_v1_initiatives__initiative_id__report_data_get",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Report Data Endpoint",
        "tags": [
          "reports"
        ]
      },
      "post": {
        "description": "Generate and return structured JSON report data for the React /report page.",
        "operationId": "generate_report_data_endpoint_api_v1_initiatives__initiative_id__report_data_post",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Generate Report Data Endpoint",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/initiatives/{initiative_id}/report/mail": {
      "post": {
        "description": "Email the compliance report as a PDF attachment to the authenticated user.\n\nGenerates HTML on the fly from current answers so it always works,\nregardless of whether the old /report endpoint has been called.",
        "operationId": "mail_report_api_v1_initiatives__initiative_id__report_mail_post",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Mail Report",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/initiatives/{initiative_id}/report/pdf": {
      "get": {
        "description": "Generate the compliance report as a PDF and return it as a file download.",
        "operationId": "download_report_pdf_api_v1_initiatives__initiative_id__report_pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Download Report Pdf",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}